pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bafle"
version = "0.0.1"
authors = [
  { name="Tomás Cabrera", email="tcabrera@andrew.cmu.edu" },
]
description = "BBH-AGN FLare Experiments"
readme = "README.md"
requires-python = ">=3.8,<3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "matplotlib",
    "astropy",
    "scipy",
    "pandas",
    "tqdm",
    "requests",
    "pytest",
    "astropy_healpix",
    "parmap",
    "ligo.skymap @ git+https://github.com/tomas-cabrera/ligo.skymap.git",
]

[project.urls]
Homepage = "https://github.com/tomas-cabrera/bafle"
Issues = "https://github.com/tomas-cabrera/bafle/issues"
